refactor: extract card list logic into useCardList custom hook
